requirements.txt
aniso8601==9.0.1
bcrypt==4.1.2
boto3==1.33.13
botocore==1.33.13
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contextlib2==21.6.0
dnspython==2.3.0
docopt==0.6.2
Flask==2.2.5
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
flask-swagger-ui==4.11.1
idna==3.6
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
pipreqs==0.4.13
PyJWT==2.8.0
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.8.2
schema==0.7.5
six==1.16.0
typing_extensions==4.7.1
urllib3==1.26.18
Werkzeug==2.2.3
yarg==0.1.9
zipp==3.15.0